Configuration for placing the xauthority file in node fs root
Added a configuration option to place the temporary xauthority file
on the fs root of the node instead of the in the build's workspace.
This can be used as a workaround for JENKINS-19139 if the path to
the fs root doesn't contain any spaces but for example the folder
to the job does.